# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sticker Matchbox API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /albums:
    get:
      operationId: listAlbums
      tags: [albums]
      summary: List all published albums
      responses:
        "200":
          description: Albums list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
    post:
      operationId: createAlbum
      tags: [albums]
      summary: Admin create album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumBody"
      responses:
        "201":
          description: Album created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /albums/{albumId}:
    get:
      operationId: getAlbum
      tags: [albums]
      summary: Get album with stickers
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Album with stickers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumWithStickers"
    put:
      operationId: updateAlbum
      tags: [albums]
      summary: Admin update album
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumBody"
      responses:
        "200":
          description: Album updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /albums/{albumId}/publish:
    patch:
      operationId: toggleAlbumPublish
      tags: [albums]
      summary: Admin publish/unpublish album
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishBody"
      responses:
        "200":
          description: Album publish state updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /albums/{albumId}/stickers:
    get:
      operationId: listAlbumStickers
      tags: [albums]
      summary: Get all stickers for an album
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stickers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sticker"
    post:
      operationId: batchInsertStickers
      tags: [albums]
      summary: Admin batch insert stickers from text list
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchStickersBody"
      responses:
        "201":
          description: Stickers inserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchStickersResponse"
  /albums/{albumId}/stickers/{stickerId}:
    put:
      operationId: updateSticker
      tags: [albums]
      summary: Admin update a sticker
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
        - name: stickerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStickerBody"
      responses:
        "200":
          description: Sticker updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sticker"

  # USER ALBUMS
...
  /user/albums/{albumId}:
    post:
      operationId: addAlbumToUser
      tags: [user-albums]
      summary: Add album to user profile
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Album added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeAlbumFromUser
      tags: [user-albums]
      summary: Remove album from user profile
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Album removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # USER STICKERS
  /user/albums/{albumId}/stickers:
    get:
      operationId: getUserAlbumStickers
      tags: [user-stickers]
      summary: Get user's sticker states for an album
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User sticker states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSticker"
  /user/albums/{albumId}/stickers/{stickerId}:
    patch:
      operationId: updateUserStickerState
      tags: [user-stickers]
      summary: Update sticker state (cycle mancante/posseduta/doppia)
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: integer
        - name: stickerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStickerStateBody"
      responses:
        "200":
          description: State updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSticker"

  # MATCHES
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Album:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        totalStickers:
          type: integer
        isPublished:
          type: boolean
        createdAt:
          type: string
      required: [id, title, totalStickers, isPublished]
    AlbumWithStickers:
      allOf:
        - $ref: "#/components/schemas/Album"
        - type: object
          properties:
            stickers:
              type: array
              items:
                $ref: "#/components/schemas/Sticker"
          required: [stickers]
    Sticker:
      type: object
      properties:
        id:
          type: integer
        albumId:
          type: integer
        number:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [id, albumId, number, name]
    CreateAlbumBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coverUrl:
          type: string
        isPublished:
          type: boolean
      required: [title]
    PublishBody:
      type: object
      properties:
        isPublished:
          type: boolean
      required: [isPublished]
    BatchStickersBody:
      type: object
      properties:
        rawList:
          type: string
          description: "Numbered list, one per line: '1 Messi' or '1. Messi'"
        stickers:
          type: array
          items:
            $ref: "#/components/schemas/UpdateStickerBody"
      required: []
    BatchStickersResponse:
      type: object
      properties:
        inserted:
          type: integer
        stickers:
          type: array
          items:
            $ref: "#/components/schemas/Sticker"
      required: [inserted, stickers]
    UpdateStickerBody:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string
        description:
          type: string
      required: [number, name]
...
    UserSticker:
      type: object
      properties:
        stickerId:
          type: integer
        albumId:
          type: integer
        state:
          type: string
          enum: [mancante, posseduta, doppia]
        number:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [stickerId, albumId, state, number, name]
    UpdateStickerStateBody:
      type: object
      properties:
        state:
          type: string
          enum: [mancante, posseduta, doppia]
      required: [state]
...