# lib/api-zod/src/generated/types/batchStickersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sticker Matchbox API
 * OpenAPI spec version: 0.1.0
 */
import type { Sticker } from "./sticker";

export interface BatchStickersResponse {
  inserted: number;
  stickers: Sticker[];
}