# lib/api-zod/src/generated/types/updateStickerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sticker Matchbox API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStickerBody {
  number: number;
  name: string;
  description?: string;
}