# lib/api-zod/src/generated/types/updateStickerStateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sticker Matchbox API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStickerStateBodyState } from "./updateStickerStateBodyState";

export interface UpdateStickerStateBody {
  state: UpdateStickerStateBodyState;
}